pyproject.toml
[tool.poetry]
name = "multirocket-ecg-ablation"
version = "0.1.0"
description = "MultiROCKET ECG classification ablation study: autoencoder vs pooling, MLP vs FT-Transformer"
readme = "README.md"
authors = []
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.24"
pandas = "^2.0"
scikit-learn = "^1.3"
pyyaml = "^6.0"
joblib = "^1.3"
torch = "^2.0"
tqdm = "^4.65"
matplotlib = "^3.7"
jupyter = "^1.0"
sktime = { version = "^0.25", optional = true }

[tool.poetry.extras]
sktime = ["sktime"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
